+++ server/src/Tgm.Roborally.Server/Controllers/MapApi.cs
@@ -14,8 +14,8 @@ using Swashbuckle.AspNetCore.Annotations;
 using Tgm.Roborally.Server.Attributes;
 using Tgm.Roborally.Server.Authentication;
 using Tgm.Roborally.Server.Engine;
+using Tgm.Roborally.Server.Engine.Exceptions;
 using Tgm.Roborally.Server.Models;
-using Map = Tgm.Roborally.Server.Models.Map;
 
 namespace Tgm.Roborally.Server.Controllers {
 	/// <summary>
@@ -36,9 +36,11 @@ namespace Tgm.Roborally.Server.Controllers {
 		[ValidateModelState]
 		[SwaggerOperation("GetGameMap")]
 		[SwaggerResponse(200, type: typeof(Map), description: "OK")]
-		public virtual IActionResult GetGameMap([FromRoute(Name = "game_id")] [Required] int gameId) =>
+		public virtual IActionResult GetGameMap([FromRoute(Name = "game_id")] [Required]
+												int gameId) =>
 			new GameRequestPipeline()
 				.Game(gameId)
+				.FailIfNull(c => c.Game.Map, GameExceptions.MAP_NOT_EXISTING)
 				.Compute(code: c => c.Game.Map.CalculateEmpty())
 				.Compute(code: c => c.SetResponse(c.Game.Map))
 				.ExecuteSecure();