+++ server/src/Tgm.Roborally.Server/Engine/Managers/MapManager.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -79,9 +80,12 @@ namespace Tgm.Roborally.Server.Engine.Managers {
 		/// <param name="name"></param>
 		/// <returns></returns>
 		public Map Get(string name) {
-			FileStream     fs     = Directory.GetFiles($"{name}{MapExtedionWithDot}")[0].OpenRead();
+			FileInfo info = Directory.GetFiles($"{name}{MapExtedionWithDot}")[0];
+			if (!info.Exists)
+				return null;
+			FileStream     fs     = info.OpenRead();
 			JsonTextReader reader = new JsonTextReader(new StreamReader(fs));
-			Map            m      = (Map) serializer.Deserialize(reader);
+			Map            m      = serializer.Deserialize<Map>(reader);
 			reader.Close();
 			fs.Close();
 			return m;
@@ -109,6 +113,8 @@ namespace Tgm.Roborally.Server.Engine.Managers {
 				}
 
 				dir.Create();
+				if (dir.FullName.Equals(Directory.FullName))
+					break;
 			}
 		}
 	}