+++ server/src/Tgm.Roborally.Server/Engine/Phases/PreExecutionPhase.cs
@@ -4,16 +4,15 @@ using Tgm.Roborally.Server.Models;
 
 namespace Tgm.Roborally.Server.Engine.Phases {
 	public class PreExecutionPhase : GamePhase {
-		private GameLogic _game;
 
 		protected override object Information => new {
-			register = _game.executionState.CurrentRegister
+			register = Game.executionState.CurrentRegister
 		};
 
 		public override GameState NewState => GameState.PLAYING;
 
 		protected override GamePhase Run(GameLogic game) {
-			_game = game;
+			Game.executionState.RestartRobotCycle();
 			return new PreStatementPhase();
 		}
 
@@ -23,7 +22,6 @@ namespace Tgm.Roborally.Server.Engine.Phases {
 
 		public override bool Notify(GenericEvent ev) => false;
 
-		public override IList<EntityEventOportunity> GetPossibleActions(int robot, int player) =>
-			throw new NotImplementedException();
+		public override IList<EntityEventOportunity> GetPossibleActions(int robot, int player) => new List<EntityEventOportunity>();
 	}
 }