+++ server/src/Tgm.Roborally.Server/Program.cs
@@ -15,7 +15,7 @@ namespace Tgm.Roborally.Server {
 	///     Program
 	/// </summary>
 	public class Program {
-		private const string VERSION = "2.10.2";
+		private const string VERSION = "2.12.6";
 
 		/// <summary>
 		///     Main